The fonction-list livewire component now changes with regard to its mode of operation. use the component in indexparfonction view.

+++ app/resources/views/livewire/fonction-list.blade.php
@@ -22,15 +22,21 @@
                         <th scope="row">{{ $fonction->id }}</th>
                         <td>{{ $fonction->fonction_libcourt }}</td>
                         <td>{{ $fonction->fonction_liblong }}</td>
-                        <td><a href="{{ route('fonctions.show', $fonction->id) }}" class="btn btn-primary btn-sm">Consulter</a></td>
-                        <td><a href="{{ route('fonctions.edit', $fonction->id) }}" class="btn btn-info btn-sm">Editer</a></td>
-                        @can('fonctions.destroy')
-                        <td>
-                            {!! Form::open(['method' => 'DELETE','route' => ['fonctions.destroy', $fonction->id],'style'=>'display:inline']) !!}
-                            {!! Form::submit('Supprimer', ['class' => 'btn btn-danger btn-sm']) !!}
-                            {!! Form::close() !!}
-                        </td>
-                        @endcan
+                        @if ($mode == 'gestion')
+                            <td><a href="{{ route('fonctions.show', $fonction->id) }}" class="btn btn-primary btn-sm">Consulter</a></td>
+                            <td><a href="{{ route('fonctions.edit', $fonction->id) }}" class="btn btn-info btn-sm">Editer</a></td>
+                            @can('fonctions.destroy')
+                            <td>
+                                {!! Form::open(['method' => 'DELETE','route' => ['fonctions.destroy', $fonction->id],'style'=>'display:inline']) !!}
+                                {!! Form::submit('Supprimer', ['class' => 'btn btn-danger btn-sm']) !!}
+                                {!! Form::close() !!}
+                            </td>
+                            @endcan
+                        @elseif($mode == 'transformation')
+                            <td><a href="{{ route('fonctions.choixmarins', $fonction->id) }}" class="btn btn-primary btn-sm">Validation collective</a></td>
+                            <td></td>
+                            <td></td>
+                        @endif
                     </tr>
                 @endforeach
             </tbody>
